azure_function/requirements.txt
azure-functions>=1.18.0
azure-functions-worker>=1.3.0
httpx>=0.25.0
asyncio